% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{activite_FC}
\alias{activite_FC}
\title{Resting heartrate frequency}
\format{
A data frame with 40 rows and 3 variables:
\describe{
  \item{freqC}{heartrate}
  \item{Sport}{Three different activities }
  \item{Sexe}{F for women and H for men}
}
}
\usage{
activite_FC
}
\description{
A dataset containing the heartrate for 24 individuals, 20 men and 20 women, use to illustrate anova model
}
\keyword{datasets}
