% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{bats}
\alias{bats}
\title{Bats analysis of brain size in relation to foraging ecology}
\format{

}
\source{
\url{https://biology.ucr.edu/people/faculty/Garland/HutcEA02.pdf}
}
\usage{
bats
}
\description{
A dataset containing information on Diets as well as morphological information on different bats species
}
\keyword{datasets}
