% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{fou}
\alias{fou}
\title{fou data}
\format{
A data frame with 5881 rows and 14 variables:
\describe{
  \item{time_step}{individual position id }
  \item{datetime}{date}
  \item{alt}{recorded altitude}
  \item{dist.nid}{distance to a central place}
  \item{ID}{bird ID}
  \item{Easting}{the x coordinates}
  \item{Northing}{the y coordinates}

}
}
\usage{
fou
}
\description{
fou a dataset containing the recorded position from  3 Brazilian red footed boobies
}
\keyword{datasets}
