% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{freqdata}
\alias{freqdata}
\title{Resting heartrate frequency}
\format{
A data frame with 40 rows and 3 variables:
\describe{
  \item{freqC}{heartrate}
  \item{Sport}{a sports practice level}
  \item{Sexe}{F for women and H for men}
}
}
\source{
\url{https://raw.githubusercontent.com/MarieEtienne/ModStat/master/data/FreqCardiaqueEq.txt}
}
\usage{
freqdata
}
\description{
A dataset containing the heartrate for 40 individuals, 20 men and 20 women, use to illustrate anova model
}
\keyword{datasets}
