% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{grenouilles}
\alias{grenouilles}
\title{California red-legged frog presence}
\format{
A data frame with 236 rows and 5 variables to the study the spatial repartition of the Calilfornia red-legged frogs.:
\describe{
  \item{Source}{primary source of the presence or absence status}
  \item{Source2}{additional information on the sources}
  \item{Status}{Presence or absence of the frogs on the site}
  \item{Latitude}{Site Latitude}
  \item{Longitude}{Site Longitude}
  }
}
\usage{
grenouilles
}
\description{
California red-legged frog presence
}
\keyword{datasets}
