% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{pins}
\alias{pins}
\title{Pine trees allometry}
\format{
A data frame with 39 rows and 3 variables:
\describe{
  \item{Espece}{Pine species (Blanc/jaune)}
  \item{Diametre}{Diameter in inches}
  \item{Hauteur}{Height in feet}
  }
}
\usage{
pins
}
\description{
Pine trees allometry
}
\keyword{datasets}
