% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{pruniers}
\alias{pruniers}
\title{Plum trees transplantation rates}
\format{
A data frame with 8 rows and 4 variables:
\describe{
  \item{Longueur}{Root length at the time of transplantation, (court ou long)}
  \item{Date}{Transplant period (automne/printemps)}
  \item{Etat}{Tree status (Vivant/Mort)}
  \item{Effectif}{Number of trees}
  }
}
\usage{
pruniers
}
\description{
Plum trees transplantation rates
}
\keyword{datasets}
