% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{usdata}
\alias{usdata}
\title{Annual average over the 1969 1971 time period of several atmospheric variables for different cities}
\format{
A data frame with 41 rows and 8 variables:
\describe{
  \item{City}{Name of the city of interest}
  \item{SO2}{Sulphur dioxide which increases the risk of acid rain}
  \item{temp}{Average temperature in Fahrenheit}
  \item{manuf}{No. of companies employing more than 20 employees }
  \item{pop}{Population in thousands}
  \item{wind}{Maverage annual wind speed in miles/hour}
  \item{precip}{annual precipitation height in inches}
  \item{days}{No. of days of precipitation}
  }
}
\usage{
usdata
}
\description{
Annual average over the 1969 1971 time period of several atmospheric variables for different cities
}
\keyword{datasets}
